LPD-92395 Move VertexAIConfiguration to the ai-hub-api module
Relocate VertexAIConfiguration from ai-hub-impl to ai-hub-api and export the package so it can be reused outside ai-hub-impl. The OCD id is preserved to keep existing saved configurations valid.
