+++ modules/dxp/apps/ai-hub/ai-hub-api/bnd.bnd
@@ -3,5 +3,6 @@ Bundle-SymbolicName: com.liferay.ai.hub.api
 Bundle-Version: 2.0.0
 Export-Package:\
 	com.liferay.ai.hub.agent,\
+	com.liferay.ai.hub.configuration,\
 	com.liferay.ai.hub.guardrail,\
 	com.liferay.ai.hub.util
