+++ modules/dxp/apps/ai-hub/ai-hub-api/src/main/java/com/liferay/ai/hub/configuration/VertexAIConfiguration.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.hub.internal.configuration;
+package com.liferay.ai.hub.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -17,7 +17,7 @@ import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClass
 	scope = ExtendedObjectClassDefinition.Scope.COMPANY
 )
 @Meta.OCD(
-	id = "com.liferay.ai.hub.internal.configuration.VertexAIConfiguration",
+	id = "com.liferay.ai.hub.configuration.VertexAIConfiguration",
 	localization = "content/Language", name = "vertex-ai-configuration-name"
 )
 public interface VertexAIConfiguration {
