+++ modules/dxp/apps/ai-hub/ai-hub-api/src/main/resources/com/liferay/ai/hub/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
